requirements.txt
beautifulsoup4
requests
tabula-py
pandas
skimpy
unidecode
ipykernel